// backend-web/src/app/shared/model/produit-simple/produit-simple.model.ts
import { Annotation } from "./annotation.model";
import { CritereCalculable } from "./critere-calculable.model";
import { Distinction } from "./distinction.model";
import { Expedition } from "./expedition.model";
import { GrilleTarifaire } from "./grille-tarifaire.model";
import { IndFraisAdd } from "./ind-frais-add.model";
import { IndPromo } from "./ind-promo.model";
import { InstCaractVar } from "./inst-caract-var.model";
import { InstCaract } from "./inst-caract.model";
import { InstClassif } from "./inst-classif.model";
import { InstCommentInt } from "./inst-comment-int.model";
import { Marque } from "./marque.model";
import { MonteeEnGamme } from "./montee-en-gamme.model";
import { TarifSaisonnier } from "./tarif-saisonnier.model";
import { TarifUnitVar } from "./tarif-unit-var.model";

export class ProduitSimple {
  _id?: string;
  refProduit?: string;
  refCB?: string;
  refQR?: string;
  typePS?: string;
  nouveau?: Boolean;
  designation?: string;
  slogan?: string;
  madeIn?: string;
  descriptif?: string;
  resumeDescriptif?: string;
  imgPD?: string[];
  tag?: string;
  virtuel?: Boolean;
  telechargeable?: Boolean;
  urlExterne?: string;
  msgBtnComExterne?: string;
  affPublique?: Boolean;
  affTarif?: Number;
  venteSeule?: Boolean;
  typeVente?: string;
  uniteProduit?: string;
  minCommande?: Number;
  maxCommande?: Number;
  typeTarif?: Number;
  tarifUHT?: Number;
  tVA?: Number;
  pays?: string;
  tarVarParSaison?: Boolean;
  tarVarParZoneGeo?: Boolean;
  tarVarParDsbtr?: Boolean;
  tarVarParVrte?: Boolean;
  tarVarParQte?: Boolean;
  notation?: Number;
  affAvis?: Boolean;
  etatPub?: string;
  msgAchat?: string;
  ordreAff?: Number;
  gestStock?: Boolean;
  indEtatStock?: Boolean;
  expedition?: Expedition;
  annotations?: Annotation[];
  criteresCalculables?: CritereCalculable[];
  distinctions?: Distinction[];
  grilleTarifaires?: GrilleTarifaire[];
  instCaracterists?: InstCaract[];
  indicationPromo?: IndPromo[];
  indicationFraisAdd?: IndFraisAdd[];
  insCaractVar?: InstCaractVar;
  instClassifs?: InstClassif[];
  instCommentaireInterne?: InstCommentInt;
  marque?: Marque;
  monteeEnGamme?: MonteeEnGamme[];
  tarifSaisonnier?: Array<any>;
  tarifUV?: TarifUnitVar;
}